allow removal of goals
